hello world cleanup + updates
