+++ website/versioned_docs/version-0.28/DOC_RENATIVE_RUNTIME.md
@@ -133,12 +133,12 @@ import { isTizenwatch } from 'renative'
 isTizenwatch()
 ```
 
-### isTizenphone
+### isTizenmobile
 
 ```js
-import { isTizenphone } from 'renative'
+import { isTizenmobile } from 'renative'
 
-isTizenphone()
+isTizenmobile()
 ```
 
 ### isMacos
